edit runlist tabpane
